+++ src/leiningen/new/war.clj
@@ -18,8 +18,7 @@
   (if (some #{"+war"} (:features options))
     [(update-assets assets)
      (-> options
-         (rename-keys {:http-server-dependencies :dev-http-server-dependencies})
-         (dissoc :immutant-session)
+         (dissoc :immutant-session :http-server-dependencies)
          (assoc
            :war true
            :uberwar-options (indent root-indent (ring-options options)))
@@ -28,7 +27,8 @@
                                    dependencies
                                    (conj dependencies ['luminus/ring-ttl-session "0.3.1"]))))
          (append-options :dependencies [['ring/ring-servlet "1.4.0"]])
-         (append-options :dev-dependencies [['directory-naming/naming-java "0.8"]])
+         (append-options :dev-dependencies (into [['directory-naming/naming-java "0.8"]]
+                                                 (:http-server-dependencies options)))
          (append-options :plugins (if (some #{"+lein"} (:features options))
                                     [['lein-uberwar "0.2.0"]])))]
     state))